pyproject.toml
[tool.poetry]
name = "nonebot_plugin_groupAccountant"
version = "0.0.2"
description = "A group chat accountant for nonebot"
authors = ["shilapi <shilapi@outlook.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/shilapi/nonebot-plugin-groupAccountant"
repository = "https://github.com/shilapi/nonebot-plugin-groupAccountant"

[tool.poetry.dependencies]
python = "^3.10.6"
nonebot2 = "^2.0.1"
nonebot-adapter-onebot = "^2.2.4"
nonebot_plugin_localstore = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"